requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
docutils==0.21.2
fastapi==0.109.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.8
ifaddr==0.2.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.2.0
markdown2==2.5.0
MarkupSafe==2.1.5
multidict==6.0.5
newspaper4k==0.9.3.1
nicegui==1.4.36
nltk==3.9.1
numpy==2.1.0
orjson==3.10.7
pandas==2.2.2
pillow==10.4.0
pscript==0.7.7
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.36.3
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vbuild==0.8.2
watchfiles==0.23.0
websockets==13.0
wsproto==1.2.0
yarl==1.9.4
